scripts/requirements.txt
pyjwt==2.9.0
websockets==12.0
